In this game, the white ball which is placed in a board will move according to the commands and collect points according collision with balls of different colors. The user plays the game by moving the white ball denoted by character ’*’ to horizontally or vertically adjacent cells.
The user earns/lose points if he/she move the
white ball to a cell containing balls of Red (R), Yellow (Y), and Black (B). If the user move
the white ball to a cell containing a Hole (H), the game is over since the ball will fall in the
this hole.